+++ packages/backend/src/nest/connections-manager/connections-manager.service.ts
@@ -30,11 +30,8 @@ import {
   NetworkStats,
   PushNotificationPayload,
   RegisterOwnerCertificatePayload,
-  RegisterUserCertificatePayload,
   RemoveDownloadStatus,
   ResponseCreateNetworkPayload,
-  SaveCertificatePayload,
-  SaveOwnerCertificatePayload,
   SendCertificatesResponse,
   SendMessagePayload,
   SetChannelSubscribedPayload,
@@ -43,7 +40,6 @@ import {
   UploadFilePayload,
   PeerId as PeerIdType,
   SaveCSRPayload,
-  SendUserCertificatePayload,
   CommunityMetadata,
   CommunityMetadataPayload,
 } from '@quiet/types'
@@ -177,10 +173,6 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
   public async closeAllServices(options: { saveTor: boolean } = { saveTor: false }) {
     if (this.tor && !options.saveTor) {
       await this.tor.kill()
-    }
-    if (this.registrationService) {
-      this.logger('Stopping registration service')
-      await this.registrationService.stop()
     }
     if (this.storageService) {
       this.logger('Stopping orbitdb')
@@ -395,22 +387,17 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
     this.socketService.on(SocketActionTypes.CONNECTION_PROCESS_INFO, data => {
       this.serverIoProvider.io.emit(SocketActionTypes.CONNECTION_PROCESS_INFO, data)
     })
-
-    this.registrationService.on(SocketActionTypes.CONNECTION_PROCESS_INFO, data => {
-      this.serverIoProvider.io.emit(SocketActionTypes.CONNECTION_PROCESS_INFO, data)
-    })
   }
   private attachRegistrationListeners() {
-    this.registrationService.on(RegistrationEvents.REGISTRAR_STATE, (payload: ServiceState) => {
-      this.registrarState = payload
-    })
     this.registrationService.on(SocketActionTypes.SAVED_OWNER_CERTIFICATE, payload => {
       this.serverIoProvider.io.emit(SocketActionTypes.SAVED_OWNER_CERTIFICATE, payload)
     })
+    // Not used atm
     this.registrationService.on(RegistrationEvents.ERROR, payload => {
       emitError(this.serverIoProvider.io, payload)
     })
     this.registrationService.on(RegistrationEvents.NEW_USER, async payload => {
+      console.log('conncetions manager payload ', payload)
       await this.storageService?.saveCertificate(payload)
     })
   }
@@ -520,7 +507,6 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
     })
     this.storageService.on(StorageEvents.LOAD_CERTIFICATES, (payload: SendCertificatesResponse) => {
       this.serverIoProvider.io.emit(SocketActionTypes.RESPONSE_GET_CERTIFICATES, payload)
-      this.registrationService.emit(RegistrationEvents.SET_CERTIFICATES, payload.certificates)
     })
     this.storageService.on(StorageEvents.LOAD_PUBLIC_CHANNELS, (payload: ChannelsReplicatedPayload) => {
       this.serverIoProvider.io.emit(SocketActionTypes.CHANNELS_REPLICATED, payload)
@@ -575,11 +561,14 @@ export class ConnectionsManagerService extends EventEmitter implements OnModuleI
       console.log('emitting deleted channel event back to state manager')
       this.serverIoProvider.io.emit(SocketActionTypes.CHANNEL_DELETION_RESPONSE, payload)
     })
-    this.storageService.on(StorageEvents.REPLICATED_CSR, async (payload: string[]) => {
-      console.log(`On ${StorageEvents.REPLICATED_CSR}`)
-      this.serverIoProvider.io.emit(SocketActionTypes.RESPONSE_GET_CSRS, { csrs: payload })
-      payload.forEach(csr => this.registrationService.emit(RegistrationEvents.REGISTER_USER_CERTIFICATE, csr))
-    })
+    this.storageService.on(
+      StorageEvents.REPLICATED_CSR,
+      async (payload: { csrs: string[]; certificates: string[] }) => {
+        console.log(`On ${StorageEvents.REPLICATED_CSR}`)
+        this.serverIoProvider.io.emit(SocketActionTypes.RESPONSE_GET_CSRS, { csrs: payload.csrs })
+        this.registrationService.emit(RegistrationEvents.REGISTER_USER_CERTIFICATE, payload)
+      }
+    )
     this.storageService.on(StorageEvents.REPLICATED_COMMUNITY_METADATA, (payload: CommunityMetadata) => {
       console.log(`On ${StorageEvents.REPLICATED_COMMUNITY_METADATA}: ${payload}`)
       const communityMetadataPayload: CommunityMetadataPayload = {