+++ packages/backend/src/nest/registration/registration.functions.ts
@@ -1,30 +1,10 @@
-import {
-  createUserCert,
-  loadCSR,
-  CertFieldsTypes,
-  getReqFieldValue,
-  keyFromCertificate,
-  parseCertificate,
-  getCertFieldValue,
-} from '@quiet/identity'
+import { createUserCert } from '@quiet/identity'
 import { IsBase64, IsNotEmpty, validate } from 'class-validator'
-import { CertificationRequest } from 'pkijs'
-import { Agent } from 'http'
-import AbortController from 'abort-controller'
-import fetch, { Response } from 'node-fetch'
-import { getUsersAddresses } from '../common/utils'
-import {
-  ErrorCodes,
-  ErrorMessages,
-  ErrorPayload,
-  PermsData,
-  SocketActionTypes,
-  SuccessfullRegistrarionResponse,
-  UserCertificatePayload,
-  UserData,
-} from '@quiet/types'
+import { ErrorPayload, PermsData, SocketActionTypes, SuccessfullRegistrarionResponse } from '@quiet/types'
 import { CsrContainsFields, IsCsr } from './registration.validators'
 import { RegistrationEvents } from './registration.types'
+import { loadCSR, CertFieldsTypes, getCertFieldValue, getReqFieldValue, parseCertificate } from '@quiet/identity'
+import { CertificationRequest } from 'pkijs'
 import Logger from '../common/logger'
 
 const logger = Logger('registration.functions')
@@ -37,51 +17,8 @@ class UserCsrData {
 }
 
 export interface RegistrarResponse {
-  status: number
-  body: any
-}
-
-// REFACTORING: Move this method to identity package
-export const pubKeyMatch = (cert: string, parsedCsr: CertificationRequest): boolean => {
-  const parsedCertificate = parseCertificate(cert)
-  const pubKey = keyFromCertificate(parsedCertificate)
-  const pubKeyCsr = keyFromCertificate(parsedCsr)
-
-  if (pubKey === pubKeyCsr) {
-    return true
-  }
-  return false
-}
-
-export const registerOwner = async (userCsr: string, permsData: PermsData): Promise<string> => {
-  const userData = new UserCsrData()
-  userData.csr = userCsr
-  const validationErrors = await validate(userData)
-  if (validationErrors.length > 0) {
-    throw new Error(`Validation errors: ${validationErrors}`)
-  }
-  const userCert = await createUserCert(
-    permsData.certificate,
-    permsData.privKey,
-    userCsr,
-    new Date(),
-    new Date(2030, 1, 1)
-  )
-  return userCert.userCertString
-}
-
-const certificateByUsername = (username: string, certificates: string[]): string | null => {
-  /**
-   * Check if given username is already in use
-   */
-  for (const cert of certificates) {
-    const parsedCert = parseCertificate(cert)
-    const certUsername = getCertFieldValue(parsedCert, CertFieldsTypes.nickName)
-    if (certUsername?.localeCompare(username, undefined, { sensitivity: 'base' }) === 0) {
-      return cert
-    }
-  }
-  return null
+  cert: string | null
+  error: any
 }
 
 export interface RegistrationResponse {
@@ -89,194 +26,52 @@ export interface RegistrationResponse {
   data: ErrorPayload | SuccessfullRegistrarionResponse
 }
 
-export const sendCertificateRegistrationRequest = async (
-  serviceAddress: string,
-  userCsr: string,
-  communityId: string,
-  requestTimeout = 120000,
-  socksProxyAgent: Agent
-): Promise<RegistrationResponse> => {
-  const controller = new AbortController()
-  const timeout = setTimeout(() => {
-    controller.abort()
-  }, requestTimeout)
+export const extractPendingCsrs = async (payload: { csrs: string[], certificates: string[] }) => {
+  const certNames: string[] = []
+  const pendingNames: string[] = []
 
-  let options = {
-    method: 'POST',
-    body: JSON.stringify({ data: userCsr }),
-    headers: { 'Content-Type': 'application/json' },
-    signal: controller.signal,
+  payload.certificates.forEach(cert => {
+    const parsedCert = parseCertificate(cert)
+    const username = getCertFieldValue(parsedCert, CertFieldsTypes.nickName)
+    if (!username) return false
+    certNames.push(username)
+  })
+
+  const parsedCsrs: { [key: string]: CertificationRequest } = {}
+
+  for (const csr of payload.csrs) {
+    const parsedCsr = await loadCSR(csr)
+    parsedCsrs[csr] = parsedCsr
   }
 
-  options = Object.assign(
-    {
-      agent: socksProxyAgent,
-    },
-    options
-  )
+  const pendingCsrs = payload.csrs.filter(csr => {
+    const username = getReqFieldValue(parsedCsrs[csr], CertFieldsTypes.nickName)
 
-  let response: Response | null = null
+    if (!username) return false
+    if (certNames.includes(username)) return false
+    if (pendingNames.includes(username)) return false
+    pendingNames.push(username)
+    return true
+  })
 
-  try {
-    const start = new Date()
-    response = await fetch(`${serviceAddress}/register`, options)
-    const end = new Date()
-    const fetchTime = (end.getTime() - start.getTime()) / 1000
-    logger(`Fetched ${serviceAddress}, time: ${fetchTime}`)
-  } catch (e) {
-    logger.error(e)
-    return {
-      eventType: RegistrationEvents.ERROR,
-      data: {
-        type: SocketActionTypes.REGISTRAR,
-        code: ErrorCodes.NOT_FOUND,
-        message: ErrorMessages.REGISTRAR_NOT_FOUND,
-        community: communityId,
-      },
-    }
-  } finally {
-    clearTimeout(timeout)
-  }
-
-  switch (response?.status) {
-    case 200:
-      break
-    case 400:
-      return {
-        eventType: RegistrationEvents.ERROR,
-        data: {
-          type: SocketActionTypes.REGISTRAR,
-          code: ErrorCodes.BAD_REQUEST,
-          message: ErrorMessages.INVALID_USERNAME,
-          community: communityId,
-        },
-      }
-    case 403:
-      return {
-        eventType: RegistrationEvents.ERROR,
-        data: {
-          type: SocketActionTypes.REGISTRAR,
-          code: ErrorCodes.FORBIDDEN,
-          message: ErrorMessages.USERNAME_TAKEN,
-          community: communityId,
-        },
-      }
-    case 404:
-      return {
-        eventType: RegistrationEvents.ERROR,
-        data: {
-          type: SocketActionTypes.REGISTRAR,
-          code: ErrorCodes.NOT_FOUND,
-          message: ErrorMessages.REGISTRAR_NOT_FOUND,
-          community: communityId,
-        },
-      }
-    default:
-      logger.error(`Registrar responded with ${response?.status} "${response?.statusText}" (${communityId})`)
-      return {
-        eventType: RegistrationEvents.ERROR,
-        data: {
-          type: SocketActionTypes.REGISTRAR,
-          code: ErrorCodes.SERVER_ERROR,
-          message: ErrorMessages.REGISTRATION_FAILED,
-          community: communityId,
-        },
-      }
-  }
-
-  const registrarResponse: UserCertificatePayload = await response.json()
-
-  logger(`Sending user certificate (${communityId})`)
-  return {
-    eventType: SocketActionTypes.SEND_USER_CERTIFICATE,
-    data: {
-      communityId: communityId,
-      payload: registrarResponse,
-    },
-  }
+  return pendingCsrs
 }
 
-export const registerUser = async (
-  csr: string,
-  permsData: PermsData,
-  certificates: string[],
-  ownerCertificate: string
-): Promise<RegistrarResponse> => {
-  let cert: string
+export const validateCsr = async (csr: string) => {
   const userData = new UserCsrData()
   userData.csr = csr
   const validationErrors = await validate(userData)
-  if (validationErrors.length > 0) {
-    logger.error(`Received data is not valid: ${validationErrors.toString()}`)
-    return {
-      status: 400,
-      body: JSON.stringify(validationErrors),
-    }
-  }
-
-  const parsedCsr = await loadCSR(userData.csr)
-  const username = getReqFieldValue(parsedCsr, CertFieldsTypes.nickName)
-  if (!username) {
-    logger.error(`Could not parse certificate for field type ${CertFieldsTypes.nickName}`)
-    return {
-      // Should be internal server error code 500
-      status: 400,
-      body: null,
-    }
-  }
-  // Use map here
-  const usernameCert = certificateByUsername(username, certificates)
-  if (usernameCert) {
-    if (!pubKeyMatch(usernameCert, parsedCsr)) {
-      logger(`Username ${username} is taken`)
-      return {
-        // Should be conflict code 409
-        status: 403,
-        body: null,
-      }
-    } else {
-      logger('Requesting same CSR again')
-      cert = usernameCert
-    }
-  } else {
-    logger('username doesnt have existing cert, creating new')
-    try {
-      cert = await registerCertificate(userData.csr, permsData)
-    } catch (e) {
-      logger.error(`Something went wrong with registering user: ${e.message as string}`)
-      return {
-        // Should be internal server error code 500
-        status: 400,
-        body: null,
-      }
-    }
-  }
-
-  const allUsers: UserData[] = []
-  for (const cert of certificates) {
-    const parsedCert = parseCertificate(cert)
-    const onionAddress = getCertFieldValue(parsedCert, CertFieldsTypes.commonName)
-    const peerId = getCertFieldValue(parsedCert, CertFieldsTypes.peerId)
-    const username = getCertFieldValue(parsedCert, CertFieldsTypes.nickName)
-    const dmPublicKey = getCertFieldValue(parsedCert, CertFieldsTypes.dmPublicKey)
-    if (!onionAddress || !peerId || !username || !dmPublicKey) continue
-    allUsers.push({ onionAddress, peerId, username, dmPublicKey })
-  }
-
-  const peerList = await getUsersAddresses(allUsers)
-
-  return {
-    status: 200,
-    body: {
-      certificate: cert,
-      peers: peerList,
-      rootCa: permsData.certificate,
-      ownerCert: ownerCertificate,
-    },
-  }
+  return validationErrors
 }
 
-export const registerCertificate = async (userCsr: string, permsData: PermsData): Promise<string> => {
+export const issueCertificate = async (userCsr: string, permsData: PermsData): Promise<RegistrarResponse> => {
+  const validationErrors = await validateCsr(userCsr)
+  if (validationErrors.length > 0) {
+    return {
+      cert: null,
+      error: [validationErrors],
+    }
+  }
   const userCert = await createUserCert(
     permsData.certificate,
     permsData.privKey,
@@ -284,5 +79,8 @@ export const registerCertificate = async (userCsr: string, permsData: PermsData)
     new Date(),
     new Date(2030, 1, 1)
   )
-  return userCert.userCertString
+  return {
+    cert: userCert.userCertString,
+    error: null,
+  }
 }