+++ packages/identity/src/index.ts
@@ -1,4 +1,4 @@
-import { createRootCA, type RootCA } from './generateRootCA'
+import { createRootCA, type RootCA } from './createRootCA'
 import {
   extractPubKey,
   parseCertificate,
@@ -10,10 +10,10 @@ import {
   getPubKey,
 } from './extractPubKey'
 import { verifyUserCert } from './verifyUserCertificate'
-import { verifySignature } from './verification'
+import { verifySignature } from './verifySignature'
 import { sign } from './sign'
-import { createUserCsr, type UserCsr } from './requestCertificate'
-import { createUserCert, type UserCert } from './generateUserCertificate'
+import { createUserCsr, type UserCsr } from './createUserCsr'
+import { createUserCert, type UserCert } from './createUserCert'
 import {
   loadPrivateKey,
   loadCSR,
@@ -25,6 +25,8 @@ import {
   hexStringToArrayBuffer,
   arrayBufferToHexString,
   getReqFieldValue,
+  certificateByUsername,
+  pubKeyMatch,
 } from './common'
 import configCrypto from './config'
 import {
@@ -64,6 +66,8 @@ export {
   CertFieldsTypes,
   hexStringToArrayBuffer,
   arrayBufferToHexString,
+  certificateByUsername,
+  pubKeyMatch,
 }
 export { configCrypto }
 export { setupCrypto }