+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * Map messages sent before changing username
 
+* Update registrar service to match new registration flow.
+
 * Running Chromatic tests for forked PRs
 
 [2.0.0-alpha.11]